Fix clients' positions and implement support for multiple clients' connections on-the-go

+++ ClientDemo/UdpProtocol.h
@@ -14,8 +14,8 @@ struct MovePacket
 {
 	uint8_t type;
 	uint32_t actorId;
-	float x;
-	float y;
+	uint32_t x;
+	uint32_t y;
 };
 #pragma pack(pop)
